move type declaration into types folder
