+++ src/components/List/ListItem.tsx
@@ -1,4 +1,5 @@
 import React from 'react'
+import ListItemProps from '../../types/ListItemProps';
 
 const StatItem = ({ matrix, value = 1 }) => {
   const gaudgeWidth = (value/180)*100
@@ -12,125 +13,6 @@ const StatItem = ({ matrix, value = 1 }) => {
     </div>
   )
 }
-
-type Pokemon = {
-  "number": "string",
-  "name": "string",
-  "image": "string",
-  "weight": {
-    "minimum": "string",
-    "maximum": "string"
-  },
-  "height": {
-    "minimum": "string",
-    "maximum": "string"
-  },
-  "classification": "string",
-  "types": [
-    "string"
-  ],
-  "resistant": [
-    "string"
-  ],
-  "attacks": {
-    "fast": [
-      {
-        "name": "string",
-        "type": "string",
-        "damage": "number"
-      }
-    ],
-    "special": [
-      {
-        "name": "string",
-        "type": "string",
-        "damage": "number"
-      }
-    ]
-  },
-  "weaknesses": [
-    "string"
-  ],
-  "fleeRate": "number",
-  "maxCP": "number",
-  "evolutions": [
-    {
-      "number": "string",
-      "name": "string",
-      "weight": {
-        "minimum": "string",
-        "maximum": "string"
-      },
-      "height": {
-        "minimum": "string",
-        "maximum": "string"
-      },
-      "classification": "string",
-      "types": [
-        "string"
-      ],
-      "resistant": [
-        "string"
-      ],
-      "attacks": {
-        "fast": [
-          {
-            "name": "string",
-            "type": "string",
-            "damage": "number"
-          }
-        ],
-        "special": [
-          {
-            "name": "string",
-            "type": "string",
-            "damage": "number"
-          }
-        ]
-      },
-      "weaknesses": [
-        "string"
-      ],
-      "fleeRate": "number",
-      "maxCP": "number",
-      "evolutions": [
-        {
-          "number": "string",
-          "name": "string",
-          "weight": {
-            "minimum": "string",
-            "maximum": "string"
-          },
-          "height": {
-            "minimum": "string",
-            "maximum": "string"
-          },
-          "classification": "string",
-          "types": [
-            "string"
-          ],
-          "resistant": [
-            "string"
-          ],
-          "attacks": {
-            "fast": [
-              {
-                "name": "string",
-                "type": "string",
-                "damage": "number"
-              }
-            ],
-            "special": [
-              null
-            ]
-          }
-        }
-      ]
-    }
-  ]
-}
-
-type ListItemProps = { key: String; pokemon: Pokemon; }
 
 const ListItem = ({ pokemon}: ListItemProps) => {
   console.log(pokemon)
