+++ src/types/ListItemProps.ts
@@ -1,0 +1,5 @@
+import Pokemon from './Pokemon';
+
+type ListItemProps = { key: String; pokemon: Pokemon; }
+
+export default ListItemProps
