+++ src/types/TextInputProps.ts
@@ -1,0 +1,6 @@
+type TextInputProps = {
+  value: string,
+  onChange: Function,
+}
+
+export default TextInputProps
